fix warnings

+++ matrix_generator.h
@@ -3,8 +3,8 @@
 using namespace std;
 
 void generateMatrix(double** matrix, int size) {
-  for (size_t i = 0; i < size; i++)
-    for (size_t j = 0; j < size; j++)
+	for (int i = 0; i < size; i++)
+	for (int j = 0; j < size; j++)
       matrix[i][j] = rand();
 }
